server/pyproject.toml
[project]
name = "lantern-reflect"
version = "0.1.0"
description = "Pipecat backend for the Lantern Library reflection tool — voice + AI guided contributor reflections"
requires-python = ">=3.11"
dependencies = [
    "pipecat-ai[anthropic,daily,deepgram,elevenlabs,silero]>=0.0.55",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "python-dotenv>=1.0",
    "pydantic>=2.9",
    "loguru>=0.7",
    "httpx>=0.27",
]

[project.scripts]
lantern-reflect-bot = "lantern_reflect.bot:cli"
lantern-reflect-server = "lantern_reflect.server:cli"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["lantern_reflect*"]
